+++ packages/nakama-js-test/utils.ts
@@ -14,11 +14,14 @@
  * limitations under the License.
  */
 
-import { Page } from "puppeteer";
+import { Page, Browser} from "puppeteer";
 const fs = require("fs");
 const crypto = require("crypto");
 const base64url = require("base64url");
 import global from "jest-environment-puppeteer"
+
+// automatically assigned by puppeteer + Jest
+declare var browser : Browser;
 
 // util to generate a random id.
 export function generateid(): string {