+++ CHANGELOG.md
@@ -4,11 +4,23 @@ All notable changes to this project are documented below.
 The format is based on [keep a changelog](http://keepachangelog.com/) and this project uses [semantic versioning](http://semver.org/).
 
 ### [Unreleased]
+
+### Added
+- Added purchase validation for Apple, Google and Huawei.
+- Added an `ApiOverrideOperator` to leaderboard writes.
+- Added ability to logout of a session with `sessionLogout`.
+- Added realtime party support
+- Added type definitions with the protobuf adapter.
+- Added realtime party support.
+
+### Changed
 - Removes the deprecated generalist `socket.Send` function for sending data. Use the other exported Socket methods
 such as `socket.addMatchmaker` for better type checking.
-- Added realtime party support.
-- Upgraded ts-proto dependency and shipped type definitions with the protobuf adapter.
 - Changed the return type from `socket.addMatchmaker` to a `MatchmakerTicket` rather than a `MatchmakerMatched`.
+- Changed signature of `authenticateSteam` and `linkSteam` to allow for a `sync` option.
+- Sessions that are close to expiration will now be automatically refreshed. You can configure this behavior
+via the `autoRefreshSession` parameter in the `Client` constructor and the `expiredTimespanMs` expiration buffer value on the client.
+- Allowed for importing of steam friends through `importSteamFriends`.
 
 ### [2.1.7]
 ### Fixed